\name{main2003JUL26}
\docType{data}
\alias{main2003JUL26}
\title{The Aftershock Data}
\description{
  The aftershock data of 26th July 2003 earthquake of M6.2 at the northern Miyagi-Ken Japan.
}
\usage{
  data(main2003JUL26)
}
\format{
  \code{main2003JUL26} is a data frame with 2305 observations and 9 variables
 named no., longitude, latitude, magnitude, time (from the main shock in days), depth, year, month, and day.
}
\source{
  Y.Ogata (2006)
  \emph{Computer Science Monographs, No.33, Statistical Analysis of Seismicity - updated version (SASeies2006).}
  The Institute of Statistical Mathematics.
}

\keyword{datasets}
