\name{PoissonData}
\docType{data}
\alias{PoissonData}
\title{Poisson Data}
\description{
  Poisson test data for \code{ptspec}.
}
\usage{
data(PoissonData)
}
\format{
  A numeric vector of length 2553.
}
\source{
  Ogata, Y., Katsura, K. and Zhuang, J. (2006) \emph{Computer Science Monographs, No.32,
  TIMSAC84: STATISTICAL ANALYSIS OF SERIES OF EVENTS (TIMSAC84-SASE) VERSION 2}.
  The Institute of Statistical Mathematics.
}

\keyword{datasets}
