\name{res2003JUL26}
\docType{data}
\alias{res2003JUL26}
\title{The Residual Point Process Data}
\description{
  The residual point process data of 26th July 2003 earthquake of M6.2 at the northern Miyagi-Ken Japan.
}
\usage{
  data(res2003JUL26)
}
\format{
  \code{res2003JUL26} is a data frame with 553 observations and 7 variables named 
no., longitude, latitude, magnitude, time (from the main shock in days), depth, Ft (transformed time).
}
\source{
  Y.Ogata (2006)
  \emph{Computer Science Monographs, No.33, Statistical Analysis of Seismicity - updated version (SASeies2006).}
  The Institute of Statistical Mathematics.
}

\keyword{datasets}
