\name{SAPP-package}
\alias{SAPP-package}
\alias{SAPP}
\docType{package}

\title{
Statistical Analysis of Point Processes
}
\description{
R functions for statistical analysis of point processes
}
\details{
This package provides functions for statistical analysis of series of events and
 seismicity.

For overview of point process models, 'Statistical Analysis of Point Processes
 with R' is available in the package vignette using the vignette function
 (e.g., vignette("SAPP")).

}
\references{
Ogata, Y., Katsura, K. and Zhuang, J. (2006) \emph{Computer Science Monographs,
 No.32, TIMSAC84: STATISTICAL ANALYSIS OF SERIES OF EVENTS (TIMSAC84-SASE)
 VERSION 2}. The Institute of Statistical Mathematics.
 \url{http://www.ism.ac.jp/editsec/csm/}

Ogata, Y. (2006)
\emph{Computer Science Monographs, No.33, Statistical Analysis of Seismicity -
 updated version (SASeies2006).} The Institute of Statistical Mathematics.
 \url{http://www.ism.ac.jp/editsec/csm/}
}

\keyword{ package }


