\name{SelfExcit}
\docType{data}
\alias{SelfExcit}
\title{Self-Exciting Point Process Data}
\description{
  Self-exciting point process test data for \code{linlin}.
}
\usage{
data(SelfExcit)
}
\format{
  A numeric vector of length 99.
}
\source{
  Ogata, Y., Katsura, K. and Zhuang, J. (2006) \emph{Computer Science Monographs, No.32,
  TIMSAC84: STATISTICAL ANALYSIS OF SERIES OF EVENTS (TIMSAC84-SASE) VERSION 2}.
  The Institute of Statistical Mathematics.
}

\keyword{datasets}
