\name{Brastings}
\docType{data}
\alias{Brastings}
\title{The Occurrence Times Data}
\description{
  This data consists of the occurrence times of 627 brastings at a certain stoneyard
 with very small portion of microearthquakes during a past 4600days.
}
\usage{
  data(Brastings)
}
\format{
  A numeric vector of length 627.
}
\source{
  Y.Ogata, K.Katsura and J.Zhuang (2006)
  \emph{Computer Science Monographs, No.32, TIMSAC84: Statistical Analysis of Series of events (TIMSAC84-SASE) Version 2.}
  The Institute of Statistical Mathematics.
}

\keyword{datasets}
