% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_feature.R
\name{influence_feature}
\alias{influence_feature}
\title{Compute the most influential feature per observation (Model-Agnostic)}
\usage{
influence_feature(data, model, scale_data = TRUE)
}
\arguments{
\item{data}{Data frame of features.}

\item{model}{A fitted model (\code{lm}, \code{glm}, \code{randomForest}, \code{ranger}).}

\item{scale_data}{Boolean. If TRUE, scales data for lm to match PCA (default: TRUE).}
}
\value{
A character vector of feature names (most influential per row).
}
\description{
Compute the most influential feature per observation (Model-Agnostic)
}
\examples{
\donttest{
data(iris)
X <- iris[, 1:4]
model <- lm(Petal.Width ~ Sepal.Length + Sepal.Width + Petal.Length, data = iris)
influence_feature(X, model)
}
}
