\name{PProcess}
\docType{data}
\alias{PProcess}
\title{The Point Process Data}
\description{
  The point process test data for \code{linsim} and \code{linlin}.
}
\usage{
  data(PProcess)
}
\format{
  A numeric vector of length 72.
}
\source{
  Y.Ogata, K.Katsura and J.Zhuang (2006) \emph{Computer Science Monographs, No.32,
  TIMSAC84: STATISTICAL ANALYSIS OF SERIES OF EVENTS (TIMSAC84-SASE) VERSION 2}.
  The Institute of Statistical Mathematics.
}

\keyword{datasets}
