\name{SAPP-package}
\alias{SAPP-package}
\alias{SAPP}
\docType{package}

\title{
Statistical Analysis of Point Processes
}
\description{
R functions for statistical analysis of point processes
}
\details{
This package provides functions for statistical analysis of series of events and seismicity.

For overview of point process models,  see \url{../doc/SAPP-guide_e.pdf}.
  PDF version of reference manual is  available in \url{../doc/SAPP-manual.pdf}

}
\references{
Y.Ogata, K.Katsura and J.Zhuang (2006) \emph{Computer Science Monographs, No.32,
TIMSAC84: STATISTICAL ANALYSIS OF SERIES OF EVENTS (TIMSAC84-SASE) VERSION 2}.
The Institute of Statistical Mathematics. \url{http://www.ism.ac.jp/editsec/csm/index.html}

Y.Ogata (2006)
\emph{Computer Science Monographs, No.33, Statistical Analysis of Seismicity - updated version (SASeies2006).}
The Institute of Statistical Mathematics. \url{http://www.ism.ac.jp/editsec/csm/index.html}
}

\keyword{ package }


